+++ src/widgets/MyDrafts.php
@@ -45,7 +45,7 @@ class MyDrafts extends Widget
     /**
      * @inheritdoc
      */
-    public static function icon()
+    public static function icon(): ?string
     {
         return Craft::getAlias('@appicons/draft.svg');
     }
@@ -78,7 +78,7 @@ class MyDrafts extends Widget
     /**
      * @inheritdoc
      */
-    public function getBodyHtml()
+    public function getBodyHtml(): ?string
     {
         $drafts = Entry::find()
             ->drafts()